Added core scripts used in setup/operation
- duration-test.sh -> tests how long device runs on battery
- install-setwanip.sh -> installs a script that modifies a line in the
index.html file for KA-Lite’s IP
- khanlink*py scripts -> used in moving Khan Academy videos around in
the folder structure so they are only on the disk one time each
- rpi-wiggle.sh -> expands the file system size to about 1GB less then
the total SD card size.  Useful when you want to copy the SD card to
another card the might be slightly smaller (not all SD cards are the
same size).